pyproject.toml
[project]
name = "filedrench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioftp>=0.27.2",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "fastapi>=0.120.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "toml>=0.10.2",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
]

[tool.basedpyright]
typeCheckingMode = "standard"
